Refactor: Move WorkFieldShowService TO Admin Package (#93)
